Feat: Edit contract
* feat: conclude edit contract feature

* feat: download document to docx
