+++ package.json
@@ -20,11 +20,13 @@
         "@mui/material": "^5.13.5",
         "@mui/x-data-grid": "^6.10.0",
         "@reduxjs/toolkit": "^1.9.5",
+        "@types/file-saver": "^2.0.7",
         "antd": "^5.8.5",
         "axios": "^1.4.0",
+        "file-saver": "^2.0.5",
         "firebase": "^10.1.0",
         "formik": "^2.4.2",
-        "html-docx-js": "^0.3.1",
+        "html-docx-js-typescript": "^0.1.5",
         "i18next": "^23.5.1",
         "mammoth": "^1.6.0",
         "moment": "^2.29.4",
@@ -35,6 +37,7 @@
         "react-i18next": "^13.2.2",
         "react-markdown": "^8.0.7",
         "react-multi-email": "^1.0.12",
+        "react-quill": "^2.0.0",
         "react-redux": "^8.1.1",
         "react-router-dom": "^6.13.0",
         "react-scripts": "5.0.1",
