+++ src/i18n/locales/en.json
@@ -12,6 +12,7 @@
       "delete": "Delete",
       "settings": "Settings",
       "edit": "Edit",
+      "save": "Save",
       "private": "Private",
       "public": "Public",
       "loading": "Loading",
