+++ src/i18n/locales/es.json
@@ -12,6 +12,7 @@
       "delete": "Eliminar",
       "settings": "Configuración",
       "edit": "Editar",
+      "save": "Guardar",
       "private": "Privado",
       "public": "Público",
       "loading": "Cargando",
