+++ src/i18n/locales/pt.json
@@ -12,6 +12,7 @@
       "delete": "Deletar",
       "settings": "Configurações",
       "edit": "Editar",
+      "save": "Salvar",
       "private": "Privado",
       "public": "Público",
       "loading": "Carregando",
