+++ src/pages/ContractAnalysis/index.tsx
@@ -7,15 +7,15 @@ import {
   Space,
   Tooltip,
   Tour,
-  TourProps,
   Typography,
   message,
 } from "antd";
 import { Base } from "../../layouts/Base";
 import { AnalysisContainer, ContractContainer, Page, SelectedTextContainer } from "./styled";
-import { v4 as uuidv4 } from "uuid";
 import {
   CopyOutlined,
+  DownloadOutlined,
+  EditOutlined,
   FlagOutlined,
   MessageOutlined,
   QuestionCircleOutlined,
@@ -23,19 +23,17 @@ import {
   UpOutlined,
   UserOutlined,
 } from "@ant-design/icons";
-import { useEffect, useRef, useState } from "react";
-import { useAppDispatch } from "../../redux/hooks";
+import { useEffect } from "react";
 import { ContractTool } from "../../components/Contracts/ContractTool";
 import Search from "antd/es/input/Search";
 import { IContract, contractsServices } from "../../services";
 import Paragraph from "antd/es/typography/Paragraph";
-import { actionDisplayNotification } from "../../redux/notificationSlice";
 import { useTranslation } from "react-i18next";
-import { useNavigate, useParams } from "react-router-dom";
-import { routesManager } from "../../routes/routesManager";
-import { useWindowSize } from "../../utils/useWindowSize";
 import { MenuContainer } from "../../components/Contracts/ContractTool/styled";
-// import { Export2Word } from "../../utils/exportToWord";
+import ReactQuill from "react-quill";
+import "react-quill/dist/quill.snow.css"; // Importe o estilo CSS
+import ButtonGroup from "antd/es/button/button-group";
+import { useContract } from "./useContract";
 
 export type AnalysisList = {
   id: string;
@@ -43,179 +41,38 @@ export type AnalysisList = {
   response: string;
 };
 
-const DESKTOP_WIDTH = 1000;
-
 const ContractAnalysisPage = () => {
-  const dispatch = useAppDispatch();
-  const { width } = useWindowSize();
+  const {
+    contract,
+    updateContractHtml,
+    handleDeleteQuestion,
+    handleDeleteAnalysis,
+    handleMouseUp,
+    appendBotAskReponse,
+    appendBotAnalysisResponse,
+    setAskSomething,
+    editingDocument,
+    setEditingDocument,
+    askSomething,
+    setText,
+    steps,
+    isDesktop,
+    openTutorial,
+    loadingResponse,
+    loadingQuestion,
+    setSelectedText,
+    setOpenTutorial,
+    loading,
+    firstRender,
+    selectedText,
+    ref1,
+    ref2,
+    loadingAnalysis,
+    text,
+    setContractToState,
+    handleDownloadContract,
+  } = useContract();
   const { t } = useTranslation();
-  const navigate = useNavigate();
-  const { id } = useParams<{ id: string }>();
-  const ref1 = useRef<HTMLDivElement>(null);
-  const ref2 = useRef<HTMLDivElement>(null);
-  const firstRender = useRef(false);
-
-  const [contract, setContract] = useState<IContract>();
-  const [loading, setLoading] = useState<boolean>(true);
-  const [openTutorial, setOpenTutorial] = useState<boolean>(
-    !localStorage.getItem("contract_tutorial"),
-  );
-  const [selectedText, setSelectedText] = useState<string>();
-  // analysis state
-  const [loadingAnalysis, setLoadingAnalysis] = useState<boolean>(false);
-  // Ask state
-  const [askSomething, setAskSomething] = useState<boolean>(false);
-  const [loadingQuestion, setLoadingQuestion] = useState<boolean>(false);
-
-  const isDesktop = width >= DESKTOP_WIDTH;
-  const loadingResponse = loadingAnalysis || loadingQuestion;
-
-  const analysisPrompt = `Você é advogado que defende os interesses da parte: ${contract?.represented_party}. Analise o trecho abaixo extraído de um contrato de: ${contract?.contract_type}.Identifique possíveis riscos para a parte: ${contract?.represented_party}. Caso a redação precisar ser modificada, faça essa sugestão e aponte qual a nova redação. Só sugira alterações se for estritamente necessário. Responda com o mínimo de palavras possível!\n\n${selectedText}`;
-
-  const steps: TourProps["steps"] = [
-    {
-      title: t("pages.contractAnalysis.components.tutorial.firstStep.title"),
-      description: t("pages.contractAnalysis.components.tutorial.firstStep.description"),
-      placement: "right",
-      target: () => ref1.current!,
-      nextButtonProps: {
-        children: t("pages.contractAnalysis.components.tutorial.nextBtn"),
-      },
-    },
-    {
-      title: t("pages.contractAnalysis.components.tutorial.secondStep.title"),
-      description: t("pages.contractAnalysis.components.tutorial.secondStep.description"),
-      placement: "left",
-      target: () => ref2.current!,
-      prevButtonProps: {
-        children: t("pages.contractAnalysis.components.tutorial.previousBtn"),
-      },
-      nextButtonProps: {
-        children: t("pages.contractAnalysis.components.tutorial.nextBtn"),
-      },
-    },
-  ];
-
-  const appendBotAnalysisResponse = async () => {
-    if (!selectedText || !contract) return;
-    setContract((prev) => {
-      if (!prev) return;
-      return {
-        ...prev,
-        risk_analysis: [
-          ...prev.risk_analysis,
-          { selected: selectedText, response: "|", id: uuidv4() },
-        ],
-      };
-    });
-    try {
-      await contractsServices.sendMessageStreamContract({
-        prompt: analysisPrompt,
-        callback(data) {
-          setLoadingAnalysis(true);
-          setContract((prev) => {
-            if (!prev) return;
-            const lastAnalysis = prev.risk_analysis[prev.risk_analysis.length - 1];
-            if (data.finish_reason) {
-              lastAnalysis["response"] = lastAnalysis.response.slice(0, -1);
-              setLoadingAnalysis(false);
-              return { ...prev, risk_analysis: [...(prev.risk_analysis || [])] };
-            } else {
-              lastAnalysis["response"] = lastAnalysis.response.slice(0, -1) + data.delta + "|";
-              return { ...prev, risk_analysis: [...(prev.risk_analysis || [])] };
-            }
-          });
-        },
-      });
-    } catch (err) {
-      setContract((prev) => {
-        if (!prev) return;
-        return {
-          ...prev,
-          risk_analysis: prev.risk_analysis.slice(0, -1),
-        };
-      });
-      dispatch(
-        actionDisplayNotification({
-          messages: [(err as { message: string }).message],
-        }),
-      );
-    }
-  };
-
-  const appendBotAskReponse = async (e: string) => {
-    if (!selectedText || !e || !contract) return;
-    setLoadingQuestion(true);
-    setAskSomething(false);
-    setContract((prev) => {
-      if (!prev) return;
-      return {
-        ...prev,
-        questions: [
-          ...prev.questions,
-          { selected: selectedText, question: e, response: "|", id: uuidv4() },
-        ],
-      };
-    });
-    try {
-      const questionPrompt = `Você é advogado que defende os interesses da parte: ${contract?.represented_party}. Considerando o trecho abaixo, delimitado por ---, extraído de um contrato de: ${contract?.contract_type}, responda a pergunta delimitada por ###:
-      \n\n###${e}###
-      \n\n---${selectedText}---`;
-
-      await contractsServices.sendMessageStreamContract({
-        prompt: questionPrompt,
-        callback(data) {
-          setLoadingQuestion(true);
-          setContract((prev) => {
-            if (!prev) return;
-            const lastQuestion = prev.questions[prev.questions.length - 1];
-            if (data.finish_reason) {
-              lastQuestion.response = lastQuestion.response.slice(0, -1);
-              setLoadingQuestion(false);
-              return { ...prev, questions: [...(prev.questions || [])] };
-            } else {
-              lastQuestion.response = lastQuestion.response.slice(0, -1) + data.delta + "|";
-              return { ...prev, questions: [...(prev.questions || [])] };
-            }
-          });
-        },
-      });
-    } catch (err) {
-      setContract((prev) => {
-        if (!prev) return;
-        return {
-          ...prev,
-          risk_analysis: prev.questions.slice(0, -1),
-        };
-      });
-      dispatch(
-        actionDisplayNotification({
-          messages: [(err as { message: string }).message],
-        }),
-      );
-    } finally {
-      setAskSomething(false);
-    }
-  };
-
-  // const handleDownloadContract = () => {
-  //   try {
-  //     if (!contract) return;
-  //     Export2Word("contract_container", contract.title);
-  //   } catch (err) {
-  //     dispatch(
-  //       actionDisplayNotification({
-  //         messages: [t("global.failureRequestMessage")],
-  //       }),
-  //     );
-  //   }
-  // };
-
-  const handleMouseUp = () => {
-    const selected = window.getSelection()?.toString();
-    setAskSomething(false);
-    setSelectedText(selected || "");
-  };
 
   const items: MenuProps["items"] = [
     {
@@ -277,57 +134,19 @@ const ContractAnalysisPage = () => {
     },
   ];
 
-  const handleDeleteAnalysis = (id: string) => {
-    setContract((prev) => {
-      if (!prev) return;
-      return {
-        ...prev,
-        risk_analysis: prev.risk_analysis.filter((item) => item.id !== id),
-      };
-    });
-  };
-
-  const handleDeleteQuestion = (id: string) => {
-    setContract((prev) => {
-      if (!prev) return;
-      return {
-        ...prev,
-        questions: prev.questions.filter((item) => item.id !== id),
-      };
-    });
-  };
-
-  const setContractToState = async () => {
-    try {
-      setLoading(true);
-      if (!id) {
-        navigate(routesManager.getContractsRoute());
-        return;
-      }
-      const fetchContract = await contractsServices.getContract(id);
-      setContract(fetchContract);
-    } catch (er) {
-      dispatch(
-        actionDisplayNotification({
-          severity: "warning",
-          messages: [t("global.failureRequestMessage")],
-        }),
-      );
-    } finally {
-      setLoading(false);
-    }
-  };
-
+  // First render
   useEffect(() => {
     setContractToState();
   }, []);
 
+  // Second render / every contract change
   useEffect(() => {
     if (!contract || loading) return;
     const container = document.querySelector("#contract_container") as HTMLElement;
     if (contract.html_content && container) {
       container.innerHTML = contract.html_content;
     }
+    setText(contract.html_content);
     // Ao primeiro render do contrato não será feita alterações.
     if (!firstRender.current) {
       firstRender.current = true;
@@ -339,6 +158,18 @@ const ContractAnalysisPage = () => {
     };
     contractsServices.updateContract(contract.id, contractUpdates);
   }, [contract]);
+
+  // On edit contract
+  useEffect(() => {
+    if (!editingDocument) {
+      if (!contract || loading) return;
+      const container = document.querySelector("#contract_container") as HTMLElement;
+      if (contract.html_content && container) {
+        container.innerHTML = contract.html_content;
+      }
+      setText(contract.html_content);
+    }
+  }, [editingDocument]);
 
   if (loading || !contract) {
     return (
@@ -353,6 +184,7 @@ const ContractAnalysisPage = () => {
               />
               <Skeleton active paragraph={{ rows: 1, width: ["70%"] }} />
             </Space>
+            <Skeleton.Button active />
             <Page>
               <Skeleton active paragraph={{ rows: 25 }} />
             </Page>
@@ -383,7 +215,7 @@ const ContractAnalysisPage = () => {
     <Base title="Contract Analysis">
       <FloatButton.Group shape="circle" type="primary">
         <FloatButton.BackTop type="primary" icon={<UpOutlined />} />
-        {/* <FloatButton type="primary" icon={<DownloadOutlined />} onClick={handleDownloadContract} /> */}
+        <FloatButton type="primary" icon={<DownloadOutlined />} onClick={handleDownloadContract} />
       </FloatButton.Group>
       {isDesktop && !loading ? (
         <Tour
@@ -451,22 +283,57 @@ const ContractAnalysisPage = () => {
               </Paragraph>
             )}
           </SelectedTextContainer>
+          <ButtonGroup>
+            {editingDocument ? (
+              <>
+                <Button
+                  onClick={async () => {
+                    await updateContractHtml();
+                    setEditingDocument(false);
+                  }}
+                  type="primary"
+                >
+                  {t("global.save")}
+                </Button>
+                <Button
+                  onClick={() => {
+                    setText(contract.html_content);
+                    setEditingDocument(false);
+                  }}
+                >
+                  {t("global.cancel")}
+                </Button>
+              </>
+            ) : (
+              <Button
+                type="primary"
+                icon={<EditOutlined />}
+                onClick={() => setEditingDocument(true)}
+              >
+                {t("global.edit")}
+              </Button>
+            )}
+          </ButtonGroup>
           <Page ref={ref1}>
-            <Dropdown
-              menu={{ items }}
-              trigger={["contextMenu"]}
-              open={!!selectedText && (askSomething || undefined)}
-              autoFocus
-              align={{}}
-            >
-              <div
-                id="contract_container"
-                onMouseUp={handleMouseUp}
-                onFocus={() => setSelectedText("")}
-                onDoubleClick={handleMouseUp}
-                style={{ display: "relative" }}
-              ></div>
-            </Dropdown>
+            {!editingDocument ? (
+              <Dropdown
+                menu={{ items }}
+                trigger={["contextMenu"]}
+                open={!!selectedText && (askSomething || undefined)}
+                autoFocus
+                align={{}}
+              >
+                <div
+                  id="contract_container"
+                  onMouseUp={handleMouseUp}
+                  onFocus={() => setSelectedText("")}
+                  onDoubleClick={handleMouseUp}
+                  style={{ display: "relative" }}
+                ></div>
+              </Dropdown>
+            ) : (
+              <ReactQuill theme="snow" value={text} onChange={(value) => setText(value)} />
+            )}
           </Page>
         </ContractContainer>
         <ContractTool
